+++ obot_g474/param/param_obot_g474_motor_resolute2.c
@@ -6,8 +6,8 @@
 const Param __attribute__ ((section ("flash_param"))) param_store = {
 #include "param_default.h"
 #include "param_motor.c"
-    .main_loop_param.output_encoder.cpr = pow(2,32),
-    .fast_loop_param.motor_encoder.cpr = pow(2,32),
+    .main_loop_param.output_encoder.cpr = powf(2,32),
+    .fast_loop_param.motor_encoder.cpr = powf(2,32),
     .name = "resolute",
 #ifdef PARAM_OVERRIDES
     PARAM_OVERRIDES
